+++ app/src/main/java/com/battlelancer/seriesguide/shows/overview/ShowFragment.kt
@@ -29,6 +29,7 @@ import com.battlelancer.seriesguide.billing.BillingTools
 import com.battlelancer.seriesguide.comments.TraktCommentsActivity
 import com.battlelancer.seriesguide.databinding.LayoutRatingsBinding
 import com.battlelancer.seriesguide.getSgAppContainer
+import com.battlelancer.seriesguide.lists.ManageListsDialogFragment
 import com.battlelancer.seriesguide.notifications.NotificationService
 import com.battlelancer.seriesguide.people.Credits
 import com.battlelancer.seriesguide.people.PeopleListHelper
@@ -109,6 +110,7 @@ class ShowFragment() : Fragment() {
         val buttonNotify: MaterialButton
         val buttonHidden: MaterialButton
         val buttonEditReleaseTime: Button
+        val buttonManageLists: Button
         val buttonShortcut: MaterialButton
         val buttonLanguage: Button
         val buttonTrailer: Button
@@ -148,6 +150,7 @@ class ShowFragment() : Fragment() {
             buttonNotify = view.findViewById(R.id.buttonShowNotify)
             buttonHidden = view.findViewById(R.id.buttonShowHidden)
             buttonEditReleaseTime = view.findViewById(R.id.buttonEditReleaseTime)
+            buttonManageLists = view.findViewById(R.id.buttonShowManageLists)
             buttonShortcut = view.findViewById(R.id.buttonShowShortcut)
             buttonLanguage = view.findViewById(R.id.buttonShowLanguage)
             buttonTrailer = view.findViewById(R.id.buttonShowTrailer)
@@ -197,6 +200,11 @@ class ShowFragment() : Fragment() {
                     "custom-release-time"
                 )
             }
+        }
+
+        // Manage lists button
+        binding.buttonManageLists.setOnClickListener {
+            ManageListsDialogFragment.showForShow(parentFragmentManager, showId)
         }
 
         // Edit note button